first attempt to draw a grid
