+++ labyrinth/src/Controller/GameController.php
@@ -4,23 +4,22 @@ namespace App\Controller;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use App\Entity\Game;
 
 class GameController extends AbstractController
 {
+    private $game;
+
     /**
      * @Route("/game")
      */
     public function index()
     {
-        return $this->render('game.html.twig');
-    }
+        $this->game = new Game();
 
-    public function getGrid() : array
-    {
-        $grid = array(
-            array("A","B","C"),
-            array("D","E","F"),
-        );
-        return $grid;
+        return $this->render('game.html.twig', [
+             'game_name' => 'hello',
+             'game_grid' => $this->game->getGrid()
+         ]);
     }
 }
