+++ labyrinth/src/Entity/Game.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Entity;
+
+class Game
+{
+    public function getGrid() : array
+    {
+        $grid = array(
+            array("A","B","C"),
+            array("D","E","F"),
+        );
+        return $grid;
+    }
+}
